Add files via upload
All 3 .os9 files are disk images. None of them are bootable. They are the most current working version of decode. The PP image is all source in pretty print format and the SRC image is all of the source in standard print. The decode.os9 image is all of the binaries.