Update buildSrc.B09
This code works with getHeader1, getHeader2 and getHeader3.
getHeader3 still has an issue where the record variable to the 4th type is being assigned to the 1st type. This is due to an inability of decode to differentiate between 2 types of the same size and same number of fields. Not sure there will ever be a fix for this.
getHeader4 is still broken, but because this version works for 3 out of for test modules I am uploading it now.
I removed all unnecessary code that was commented out due to changing the gap data and the type data from array variables to record arrays. The code for the array variables is no longer valid.