Update buildSrc.B09
This now works with getHeader1,2 3 and 4. With getHeader3 there is still the issue of wrong assignment, but also I have now added code that prints the gap data that is not validated, and in getHeadser3 is producing garbage output at that location. That said, other than the integer values for the field string being incorrectly identified, it is being validated in the correct spot and the DSAT overall validates.