Update instruction.B09
While identifying all variables used in every procedure, I found one I was using in instruction that I had no DIM statement for. It was supposed to be an INTEGER, but because I had not DIM'd it, it was defined by use as a REAL. I just saved 3 bytes of data memory.